Manage JClass references more robustly
FindClass just returns a local reference, but we want to keep this around for the entire lifetime of the program. The "correct" way to do this appears to be to make it a global reference.

In practice the local reference seems to stay around forever, so not much actually changes here.