Use a blocking wait sync context for Windows on the main thread
The main thread is an STA thread, required for various GUI functionality (e.g. clipboard and file dialogs).

The STA contract normally dictates that an STA thread never blocks, as it must always pump messages, otherwise it may result in a deadlock.

.NET tries to abide by this by pumping messages when waiting on an event/monitor (using CoWaitForMultipleHandles), but this ends up causing more problems than it fixes.

This is especially prevalent if a standard wait event is used to throttle the main thread (such as waiting for the next video frame to be copied over, will be done in a future low latency mode). In this case, trying to click the icon in the task manager appears to do nothing most of the time and just rings the standard windows error sound, only sometimes actually minimizing/unminimizing the window.